+++ packages/samples/headless-react/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.7...v0.5.0) (2021-02-04)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.7](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.6...v0.5.0-alpha.7) (2021-02-04)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
+# [0.5.0-alpha.6](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.5...v0.5.0-alpha.6) (2021-02-03)
+
+**Note:** Version bump only for package headless-react
+
+
+
+
+
 # [0.5.0-alpha.5](https://github.com/coveo/ui-kit/compare/v0.5.0-alpha.4...v0.5.0-alpha.5) (2021-02-02)
 
 **Note:** Version bump only for package headless-react